Extracted states to a separate class. New type to show the status of the buttons
